Move classes into separate files and add license
